[XTENSA] Add missing userspace PC in oprofile sample backtraces that start in kernel space
Also keep upper two bits of PC unchanged through backtrace rather than take them from sp (a1).
The stack pointer is usually in the same GB (same upper 2 bits) as PC, but technically
doesn't always have to be (and might not in the future, when taking full advantage of MMU v3).
Using the previous PC makes more sense as that what the hardware does when it executes a
return instruction. Adding the saved PC from pt_regs to the trace appears to have
made the oprofile results a bit better.

Cc:            Marc Gauthier <marc@tensilica.com>
Signed-off-by:  rootPete Delaney <piet@tensilica.com>