[XTENSA] Check for external registers option for SMP
Check explicitly for an attempt to use SMP on a processor that is
not configured with the external registers option (required for SMP)
rather than fail cryptically with the assembler not recognizing
the WER or RER instruction.

Also update ISS platform defconfig to turn off SMP so that it can
work on processor configurations (variants) without that option.

Signed-off-by: Marc Gauthier <marc@tensilica.com>